feat: Lv3. 피연산자 입력 기능 구현
- 정수, 실수 입력 지원.

+++ src/main/java/org/dongwon/App.java
@@ -10,10 +10,10 @@ public class App {
             try {
                 System.out.print("첫 번째 숫자를 입력하세요: ");
                 // nextInt()를 호출하면 개행문자가 남아 String을 받아 Integer로 파싱.
-                int num1 = Integer.parseInt(sc.nextLine());
+                double num1 = Double.parseDouble(sc.nextLine());
 
                 System.out.print("두 번째 숫자를 입력하세요: ");
-                int num2 = Integer.parseInt(sc.nextLine());
+                double num2 = Double.parseDouble(sc.nextLine());
 
                 System.out.print("사칙연산 기호를 입력하세요: ");
                 char operation = sc.nextLine().trim().charAt(0);
